Fix signature dtype float32 not float64

+++ signaturizer3d/infer_signature.py
@@ -10,10 +10,10 @@ class Signaturizer:
 
     def infer_from_coordinates(
         self, atoms: list[str], coordinates: list[list[float]]
-    ) -> npt.NDArray[np.float64]:  # (n, 128) dim array for n signatures
+    ) -> npt.NDArray[np.float32]:  # (n, 128) dim array for n signatures
         sig_output = self.model.get_sig4_coordinates(atoms, coordinates)
         return sig_output
 
-    def infer_from_smiles(self, smiles_list: list[str]) -> npt.NDArray[np.float64]:
+    def infer_from_smiles(self, smiles_list: list[str]) -> npt.NDArray[np.float32]:
         sig_output = self.model.get_sig4_smiles(smiles_list)
         return sig_output
